+++ src/overlay/RandomPeerSource.h
@@ -19,13 +19,14 @@ struct PeerQuery;
 class RandomPeerSource
 {
   public:
-    static PeerQuery maxFailures(int maxFailures, bool outbound);
+    static PeerQuery maxFailures(size_t maxFailures, bool outbound);
     static PeerQuery nextAttemptCutoff(PeerType peerType);
 
     explicit RandomPeerSource(PeerManager& peerManager, PeerQuery peerQuery);
 
     std::vector<PeerBareAddress>
-    getRandomPeers(int size, std::function<bool(PeerBareAddress const&)> pred);
+    getRandomPeers(size_t size,
+                   std::function<bool(PeerBareAddress const&)> pred);
 
   private:
     PeerManager& mPeerManager;
